requirements.txt

flask
torch
torch_geometric
transformers
gitpython
networkx
numpy
pandas
scikit-learn
shap
tqdm
